Added broken Change of Gravity
not working float and bug where it gets stuck

+++ Gravity/Assets/Scripts/Constants Class.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 class Constants
 {
     public const float epsilon = 0.05f;
-    public readonly float[] lines = {6f, 0f, -6f};
+    public readonly float[] lines = {6, 0, -6};
 
-    public const float line1 = 6f;
-    public const float line2 = 0f;
-    public const float line3 = -6f;
+    public const int line1 = 6;
+    public const int line2 = 0;
+    public const int line3 = -6;
 
     public const float predictedLengthOfMove = 6f;
 }
